+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/DeclarationOrderCheckTest.java
@@ -44,7 +44,7 @@ public class DeclarationOrderCheckTest
     }
 
     @Test
-    public void testDefault() throws Exception {
+    public void testDefault1() throws Exception {
 
         final String[] expected = {
             "16:5: " + getCheckMessage(MSG_ACCESS),
@@ -56,21 +56,38 @@ public class DeclarationOrderCheckTest
             "60:9: " + getCheckMessage(MSG_STATIC),
             "69:5: " + getCheckMessage(MSG_CONSTRUCTOR),
             "95:5: " + getCheckMessage(MSG_INSTANCE),
-            "107:9: " + getCheckMessage(MSG_ACCESS),
-            "115:9: " + getCheckMessage(MSG_STATIC),
-            "121:5: " + getCheckMessage(MSG_ACCESS),
-            "126:5: " + getCheckMessage(MSG_ACCESS),
-            "131:5: " + getCheckMessage(MSG_ACCESS),
-            "134:5: " + getCheckMessage(MSG_ACCESS),
-            "140:5: " + getCheckMessage(MSG_STATIC),
-            "147:9: " + getCheckMessage(MSG_ACCESS),
-            "158:9: " + getCheckMessage(MSG_STATIC),
-            "167:5: " + getCheckMessage(MSG_CONSTRUCTOR),
-            "193:5: " + getCheckMessage(MSG_INSTANCE),
-            "198:9: " + getCheckMessage(MSG_ACCESS),
         };
         verifyWithInlineConfigParser(
-                getPath("InputDeclarationOrder.java"), expected);
+                getPath("InputDeclarationOrderDefault1.java"), expected);
+    }
+
+    @Test
+    public void testDefault2() throws Exception {
+
+        final String[] expected = {
+            "20:9: " + getCheckMessage(MSG_ACCESS),
+            "28:9: " + getCheckMessage(MSG_STATIC),
+            "34:5: " + getCheckMessage(MSG_ACCESS),
+            "39:5: " + getCheckMessage(MSG_ACCESS),
+            "44:5: " + getCheckMessage(MSG_ACCESS),
+            "47:5: " + getCheckMessage(MSG_ACCESS),
+            "53:5: " + getCheckMessage(MSG_STATIC),
+            "60:9: " + getCheckMessage(MSG_ACCESS),
+            "71:9: " + getCheckMessage(MSG_STATIC),
+            "80:5: " + getCheckMessage(MSG_CONSTRUCTOR),
+            "106:5: " + getCheckMessage(MSG_INSTANCE),
+            "111:9: " + getCheckMessage(MSG_ACCESS),
+        };
+        verifyWithInlineConfigParser(
+                getPath("InputDeclarationOrderDefault2.java"), expected);
+    }
+
+    @Test
+    public void testDefault3() throws Exception {
+        final String[] expected = CommonUtil.EMPTY_STRING_ARRAY;
+        verifyWithInlineConfigParser(
+                getPath("InputDeclarationOrderDefault3.java"),
+                expected);
     }
 
     @Test
@@ -234,14 +251,6 @@ public class DeclarationOrderCheckTest
         };
         verifyWithInlineConfigParser(
                 getPath("InputDeclarationOrderAvoidDuplicatesInStaticFinalFields.java"),
-                expected);
-    }
-
-    @Test
-    public void test() throws Exception {
-        final String[] expected = CommonUtil.EMPTY_STRING_ARRAY;
-        verifyWithInlineConfigParser(
-                getPath("InputDeclarationOrder2.java"),
                 expected);
     }
