+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/declarationorder/InputDeclarationOrderDefault3.java
@@ -8,7 +8,7 @@ ignoreModifiers = (default)false
 
 package com.puppycrawl.tools.checkstyle.checks.coding.declarationorder;
 
-public class InputDeclarationOrder2 {
+public class InputDeclarationOrderDefault3 {
     private static char[] sHexChars = {
             '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
             'A', 'B', 'C', 'D', 'E', 'F' };
